# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ZEINA Hyperviseur API
  version: "1.0.0"
  description: |
    REST + WebSocket API du superviseur IoT ZEINA.

    ## Authentification
    1. POST `/v1/auth/login` retourne un access token (15 min) + un refresh token (cookie httpOnly, 7 j).
    2. Les routes protégées attendent l'access token dans `Authorization: Bearer <token>`.
    3. Quand le token expire, POST `/v1/auth/refresh` produit un nouvel access (utilise le cookie).

    ## Modèle d'autorisation
    - **Superadmin** (`is_superadmin=true`) : bypass complet.
    - **Owner du tenant** (`tenant_role=owner`) : accès implicite à tous les sites + gestion users/rôles.
    - **Membre simple** : accès aux sites où il est listé dans `site_members`, selon le rôle.
    - Chaque rôle attribue un niveau (`none`/`read`/`write`) sur 4 fonctionnalités : `dashboard`, `devices`, `rules`, `members`.

    ## Test rapide
    1. Clique « Authorize » en haut, paste l'access token (sans le préfixe Bearer).
    2. Tous les calls authentifiés vont passer.
servers:
  - url: /
    description: |
      Origine courante (relatif). Quand tu importes ce spec dans Bruno/Postman,
      remplace par `http://localhost:5173` (via nginx) ou `http://localhost:3000`
      (API directe) — sans slash final.
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: Echange email/password contre un access token + cookie de refresh.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginReq" }
            examples:
              demo: { value: { email: admin@acme.test, password: admin123 } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/LoginResp" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Renouvelle l'access token via cookie httpOnly
      security: []
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/RefreshResp" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/auth/me:
    get:
      tags: [Auth]
      summary: Profil + snapshot des permissions par site
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/MeResp" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ---------------------------------------------------------------------------
  # SITES
  # ---------------------------------------------------------------------------
...
  /v1/devices/{id}/command:
    post:
      tags: [Commands]
      summary: Envoyer une commande à un actionneur
      description: Requiert `devices:write`. Publie une commande MQTT sur le topic `cmd/{action}`.
      parameters: [{ $ref: "#/components/parameters/DeviceId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommandReq" }
            examples:
              relay_on:
                value: { action: set, payload: { state: "on" } }
              relay_off:
                value: { action: set, payload: { state: "off" } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CommandResp" } } } }

  # ---------------------------------------------------------------------------
  # DEVICE MODELS (catalogue)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DeviceId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: ID du device
...
  responses:
    Unauthorized:
      description: Token manquant ou invalide
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    BadRequest:
      description: Validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:   { type: string, example: validation }
        message: { type: string, example: "email and password are required" }
        details: { type: object, additionalProperties: true }
...
    PermissionLevel:
      type: string
      enum: [none, read, write]
    PermissionSet:
      type: object
      properties:
        dashboard: { $ref: "#/components/schemas/PermissionLevel" }
        devices:   { $ref: "#/components/schemas/PermissionLevel" }
        rules:     { $ref: "#/components/schemas/PermissionLevel" }
        members:   { $ref: "#/components/schemas/PermissionLevel" }

    # --- Auth ---------------------------------------------------------------
    LoginReq:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string, format: password }
    LoginResp:
      type: object
      properties:
        access_token: { type: string }
        expires_at:   { type: string, format: date-time }
        user:         { $ref: "#/components/schemas/UserOut" }
    RefreshResp:
      type: object
      properties:
        access_token: { type: string }
        expires_at:   { type: string, format: date-time }
    UserOut:
      type: object
      properties:
        id:            { type: string, format: uuid }
        email:         { type: string, format: email }
        tenant_role:   { type: string, enum: [owner, member] }
        is_superadmin: { type: boolean }
        tenant_id:     { type: string, format: uuid }
        full_name:     { type: string, nullable: true }
    SiteAccess:
      type: object
      properties:
        site_id:     { type: string, format: uuid }
        site_slug:   { type: string }
        site_name:   { type: string }
        role_id:     { type: string, format: uuid, nullable: true }
        role_name:   { type: string }
        permissions: { $ref: "#/components/schemas/PermissionSet" }
    MeResp:
      type: object
      properties:
        user:  { $ref: "#/components/schemas/UserOut" }
        sites:
          type: array
          items: { $ref: "#/components/schemas/SiteAccess" }

    # --- Sites --------------------------------------------------------------
...
    CommandReq:
      type: object
      required: [action]
      properties:
        action:  { type: string, example: set }
        payload: { type: object, additionalProperties: true }
    CommandResp:
      type: object
      properties:
        command_id: { type: string, format: uuid }
        topic:      { type: string }
        issued_at:  { type: string, format: date-time }

    # --- Device models ------------------------------------------------------
...